fetching only a shallow clone of gitlab
This helps me getting through my proxy.  The proxy is really annoying.
It is too dumb to speak TLS properly and HTTPS connections break after a
while. So long transfers are not good. Shallow clones are thus better
for me.  This should work well with whatever this ansible role requires,
especially updating the Gitlab version.